Fix: pacemaker_remote: allow remote clients to timeout/reconnect
