+++ include/crm_internal.h
@@ -203,6 +203,7 @@ xmlNode *crm_remote_parse_buffer(crm_remote_t * remote);
 int crm_remote_tcp_connect(const char *host, int port);
 int crm_remote_tcp_connect_async(const char *host, int port, int timeout,       /*ms */
                                  int *timer_id, void *userdata, void (*callback) (void *userdata, int sock));
+int crm_remote_accept(int ssock);
 
 #  ifdef HAVE_GNUTLS_GNUTLS_H
 /*!
